requirements.txt
Django==5.0.4
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow
